+++ core/service/rsi/aftersales_service.go
@@ -36,7 +36,7 @@ func NewAfterSalesService(rep afterSales.IAfterSalesRepo,
 
 // 提交售后单
 func (a *afterSalesService) SubmitAfterSalesOrder(orderId int64, asType int,
-	snapshotId int32, quantity int32, reason string, imgUrl string) (int32, error) {
+	snapshotId int64, quantity int32, reason string, imgUrl string) (int32, error) {
 	ro := a._rep.CreateAfterSalesOrder(&afterSales.AfterSalesOrder{
 		// 订单编号
 		OrderId: orderId,
